[#84] Presumably fixes Babel/Webpack setup issue on Windows
The issue origin was traced down to `babel-plugin-css-modules-transform`
dependency, which has not been actively maintained for a while. As the
best solution, the library was forked and revamped to the present day
dependencies and standards, and this should fix the issue. However, it
has not been tested on Windows yet, thus there is a chance that further
changes will be required.