[Default ESLint conifg]: Disables no-plusplus rule
It turns out, ESLint rationale for this loop is that automatic semicolon
insertion may affect expressions with unary "++" and "--" operators, if
they are written in multiple lines. Such use is actually prohibited by
other rules, requesting these operators to be connected to a variable,
and end any line with explicit semicolon, thus "no-plusplus" makes no
sense in our config.