BREAKING [#139][#154] v1.1.0: <MetaTags> updates
- BREAKING [#154]: Removes automatic pre-fixing of "image" property by
website domain taken from "domain" path of the global state. It was not
a flexible solution, it is better to let end user to figure out the
appropriate way to get the domain, and provide the complete absolute
thumbnail URL into "image" property. That's how it works now.

- Adds "children" property, allowing to render a React sub-tree within
<MetaTags> node. In this case, the child sub-tree is wrapped into
the context providing this <MetaTags> properties to the components in
children hierarchy, thus allowing to easily modify title, description,
and other metatags by nested <MetaTags> (it was possible before, but
there was no convenient way to get the title, description and other
values of the parent <MetaTags> instance). The created context is
accessible via "MetaTags.Context" field.