v1.3.6: [#168] The mapping of Webpack-generated assets
Updates the build script to output the full mapping of original and
generated (emitted by Webpack) asset paths. The mapping is added to the
__asset_map__.json file in the build output folder (which replaces and
augments the __stats__.json output before). The SSR render code is
updated to provide the asset_map to SSR logic via the
ssrContext.assetMap (which can be accessed during the server-side
rendering via the react-global-state's getContext(..) hook).