[#182] Upgrades <CodeSplit> component to faciliate its use
Namely, a special `resolveRequire` function is passed now as an argument
into `getComponentServer` prop (if that prop is provided). It allows to
simplify the typical `getComponentServer` code.

Before this change, it usually had to be done this way, to prevent
Webpack from bundling in server-side only, code, or the required module,
while still being able to resolve it properly:
```
getComponentServer={
  () => {
    const path = webpack.requireWeak('path');
    const p = webpack.resolveWeak('path/to/SampleComponent');
    return webpack.requireWeap(path.resolve(__dirname, p));
  }
}
```

After this update, the same code can be simplified to such one-liner:
```
getComponentServer={
  (rr) => rr(__dirname, 'path/to/SampleComponent')
}
```